endpoint createUser e organização
