+++ semana17/projeto/src/connection.ts
@@ -1,0 +1,15 @@
+import dotenv from "dotenv";
+import knex from "knex";
+
+dotenv.config();
+
+export const connection = knex({
+	client: "mysql",
+	connection: {
+    host: process.env.DB_HOST,
+    port: 3306,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASS,
+    database: process.env.DB_NAME
+}
+});
